Add linter to CI (#26)
* Small cleanup (#24)

* bump makefile with help section

* update with linter configuration

* countOnly is unused `make lint`

* add config file and full linter options

* Add lint step to CI

* Fix/address linter findings

+++ pkg/manager/client.go
@@ -33,7 +33,7 @@ func (c *Client) members(ctx context.Context) (map[string]*Member, error) {
 	members := make(map[string]*Member)
 	for _, member := range resp.Members {
 		m := &Member{
-			ID:   uint64(member.ID),
+			ID:   member.ID,
 			Name: member.Name,
 		}
 		if len(member.ClientURLs) > 0 {
@@ -55,7 +55,7 @@ func (c *Client) addMember(ctx context.Context, peerURL string) (*Member, error)
 		return nil, err
 	}
 	m := &Member{
-		ID:   uint64(resp.Member.ID),
+		ID:   resp.Member.ID,
 		Name: resp.Member.Name,
 	}
 	if len(resp.Member.ClientURLs) > 0 {
